requirements.txt
openai~=1.93.0
Flask~=3.1.1
Werkzeug~=3.1.3
python-dotenv==1.0.0
langchain==0.0.235
faiss-cpu==1.7.4
pypdf==3.10.0

PyMySQL~=1.1.1